soc: esp32c6: Correct linker script
Previously the esp32c6 default.ld linker script was based on the esp32c3,
resulting in incorrect section addresses for internal SRAM, etc. Add
some work to correct this, based on
components/esp_system/ld/esp32c6/memory.ld.in in the ESP-IDF repo.

N.b. this linker script won't work as-in, i.e. the build will fail to
link. Committing to allow this to be pushed for review more widely.

Signed-off-by: Andrew Featherstone <andrew.featherstone@gmail.com>